+++ src/user/Admin.java
@@ -6,6 +6,14 @@ public class Admin extends User { // Inheritance
         super(username, password);
     } // Constructor
 
-    
+    public Admin(int ID, String username, String password, String email, boolean status) {
+        super(ID, username, password, email, status);
+
+    }
+
+    public Admin(String username, String password, String role, String email, boolean status) {
+        super(username, password, role, email, status);
+
+    }
 
 }