+++ src/user/Student.java
@@ -9,11 +9,9 @@ public class Student extends User{
     // Constructor for Table Functions
     public Student(int ID, String username, String password, String email, boolean status) {
         super(ID, username, password, email, status);
-
     }
 
     public Student(String username, String password, String role, String email, boolean status) {
         super(username, password, role, email, status);
-
     }
 }